+++ store/cassandra/db.go
@@ -22,6 +22,7 @@ import (
 	"errors"
 	"github.com/gocql/gocql"
 	"github.com/goph/emperror"
+	"github.com/xmidt-org/argus/model"
 	"github.com/xmidt-org/argus/store"
 	"github.com/xmidt-org/webpa-common/logging"
 	"time"
@@ -170,7 +171,7 @@ func CreateCassandraClient(config CassandraConfig, measures Measures, logger log
 	}, nil
 }
 
-func (s *CassandraClient) Push(key store.Key, item store.Item) error {
+func (s *CassandraClient) Push(key model.Key, item store.InternalItem) error {
 	err := s.client.Push(key, item)
 	if err != nil {
 		s.measures.SQLQueryFailureCount.With(store.TypeLabel, store.InsertType).Add(1.0)
@@ -180,7 +181,7 @@ func (s *CassandraClient) Push(key store.Key, item store.Item) error {
 	return nil
 }
 
-func (s *CassandraClient) Get(key store.Key) (store.Item, error) {
+func (s *CassandraClient) Get(key model.Key) (store.InternalItem, error) {
 	item, err := s.client.Get(key)
 	if err != nil {
 		if err == noDataResponse {
@@ -193,7 +194,7 @@ func (s *CassandraClient) Get(key store.Key) (store.Item, error) {
 	return item, nil
 }
 
-func (s *CassandraClient) Delete(key store.Key) (store.Item, error) {
+func (s *CassandraClient) Delete(key model.Key) (store.InternalItem, error) {
 	item, err := s.client.Delete(key)
 	if err != nil {
 		if err == noDataResponse {
@@ -206,11 +207,11 @@ func (s *CassandraClient) Delete(key store.Key) (store.Item, error) {
 	return item, err
 }
 
-func (s *CassandraClient) GetAll(bucket string) (map[string]store.Item, error) {
+func (s *CassandraClient) GetAll(bucket string) (map[string]store.InternalItem, error) {
 	item, err := s.client.GetAll(bucket)
 	if err != nil {
 		if err == noDataResponse {
-			return item, store.KeyNotFoundError{Key: store.Key{
+			return item, store.KeyNotFoundError{Key: model.Key{
 				Bucket: bucket,
 			}}
 		}