+++ store/storetest/storetest.go
@@ -19,22 +19,27 @@ package storetest
 
 import (
 	"github.com/stretchr/testify/assert"
+	"github.com/xmidt-org/argus/model"
 	"github.com/xmidt-org/argus/store"
 	"testing"
 	"time"
 )
 
-var GenericTestKeyPair = store.KeyItemPair{
-	Key: store.Key{
+var GenericTestKeyPair = store.KeyItemPairRequest{
+	Key: model.Key{
 		Bucket: "world",
 		ID:     "earth",
 	},
-	Item: store.Item{
-		Attributes: nil,
-		Value: map[string]interface{}{
-			"year":  1967,
-			"words": []string{"What", "a", "Wonderful", "World"},
+	InternalItem: store.InternalItem{
+		Item: model.Item{
+			Identifier: "earth",
+			Data: map[string]interface{}{
+				"year":  float64(1967),
+				"words": []interface{}{"What", "a", "Wonderful", "World"},
+			},
+			TTL: 3,
 		},
+		Owner: "Louis Armstrong",
 	},
 }
 
@@ -42,34 +47,34 @@ func StoreTest(s store.S, storeTiming time.Duration, t *testing.T) {
 	assert := assert.New(t)
 
 	t.Log("Basic Test")
-	err := s.Push(GenericTestKeyPair.Key, GenericTestKeyPair.Item)
+	err := s.Push(GenericTestKeyPair.Key, GenericTestKeyPair.InternalItem)
 	assert.NoError(err)
 	retVal, err := s.Get(GenericTestKeyPair.Key)
 	assert.NoError(err)
-	assert.Equal(GenericTestKeyPair.Item, retVal)
+	assert.Equal(GenericTestKeyPair.InternalItem, retVal)
 
 	items, err := s.GetAll("world")
 	assert.NoError(err)
-	assert.Equal(map[string]store.Item{"earth": GenericTestKeyPair.Item}, items)
+	assert.Equal(map[string]store.InternalItem{"earth": GenericTestKeyPair.InternalItem}, items)
 
 	retVal, err = s.Delete(GenericTestKeyPair.Key)
 	assert.NoError(err)
-	assert.Equal(GenericTestKeyPair.Item, retVal)
+	assert.Equal(GenericTestKeyPair.InternalItem, retVal)
 
 	items, err = s.GetAll("world")
 	assert.NoError(err)
-	assert.Equal(map[string]store.Item{}, items)
+	assert.Equal(map[string]store.InternalItem{}, items)
 
 	if storeTiming > 0 {
 		t.Log("staring duration tests")
-		err := s.Push(GenericTestKeyPair.Key, GenericTestKeyPair.Item)
+		err := s.Push(GenericTestKeyPair.Key, GenericTestKeyPair.InternalItem)
 		assert.NoError(err)
 		retVal, err := s.Get(GenericTestKeyPair.Key)
 		assert.NoError(err)
-		assert.Equal(GenericTestKeyPair.Item, retVal)
+		assert.Equal(GenericTestKeyPair.InternalItem, retVal)
 		time.Sleep(storeTiming + time.Second)
 		retVal, err = s.Get(GenericTestKeyPair.Key)
-		assert.Equal(store.Item{}, retVal)
+		assert.Equal(store.InternalItem{}, retVal)
 		assert.Equal(store.KeyNotFoundError{Key: GenericTestKeyPair.Key}, err)
 	}
 }