+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- added cache for webhookclient
+- reworked api
+- removed cache from webhookclient package
 
 ## [v0.1.1]
 - fixed import error with webhookclient package